refactor: fix dependency issues
